Fix a set of warnings
Warning were reported as:
warning: passing 'J_COLOR_SPACE' chooses 'int' over 'unsigned int'
warning:   in call to 'std::basic_ostream<_CharT, _Traits>& std::basic_ostream<_CharT, _Traits>::operator<<(int) [with _CharT = char, _Traits = std::char_traits<char>]'

+++ Testing/Source/InformationObjectDefinition/Cxx/TestDefs.cxx
@@ -48,9 +48,9 @@ int TestDefs(int, char *[])
         && mst != MediaStorage::CSANonImageStorage
         && mst != MediaStorage::GeneralElectricMagneticResonanceImageStorage )
         {
-        std::cerr << "Missing iod for MS: " << mst << " " <<
+        std::cerr << "Missing iod for MS: " << (int)mst << " " <<
           gdcm::MediaStorage::GetMSString(mst) << "  "; //std::endl;
-        std::cerr << "MediaStorage is " << mst << " [" << uid.GetName() << "]" << std::endl;
+        std::cerr << "MediaStorage is " << (int)mst << " [" << uid.GetName() << "]" << std::endl;
         ++ret;
         }
       }
